requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
geopandas>=0.13.0
plotly>=5.15.0
numpy>=1.24.0
pillow>=9.5.0